requirements.txt
torch~=2.3.1
monai~=1.3.0
antspyx~=0.4.2
nibabel~=5.2.1
